+++ notebooks/noj_book/known_issues.clj
@@ -29,6 +29,7 @@
 ;;
 ;; This can be used for reference or as a starting point to use Noj, if one desires.
 ;; ### scicloj.ml.smile
-;; Some of the Smile based models require native dependencies. This is documented in [Smile](https://github.com/haifengl/smile)
+;; Noj is often used together with `scicloj.ml.smile`, and
+;; some of the Smile based models require native dependencies. This is documented in [Smile](https://github.com/haifengl/smile)