+++ notebooks/noj_book/recommended_libraries.clj
@@ -11,6 +11,7 @@
 ;; * [fitdistr](https://github.com/generateme/fitdistr) - fitting distributions
 ;; * [inferme](https://github.com/generateme/inferme) - Bayesian inference
 ;; * [cmdstan-clj](https://github.com/scicloj/cmdstan-clj) - bindings for [Stan](https://mc-stan.org/) Bayesian inference - 🛠 early stage
+;; * [scicloj.ml.smile](https://github.com/scicloj/scicloj.ml.smile) - [Smile](https://haifengl.github.io/) (v 2.6) machine learning models
 ;; * [fastmath-clustering](https://generateme.github.io/fastmath-clustering/) - a wrapper for [Smile](https://haifengl.github.io/) 2.6 clustering algorithms
 ;; * [clj-djl](https://github.com/scicloj/clj-djl) - bindings for [Deep Java Library](https://djl.ai/)
 ;; * [scicloj.ml.clj-djl](https://github.com/scicloj/scicloj.ml.clj-djl) - clj-djl models for [metamorph.ml](https://github.com/scicloj/metamorph.ml)